qrtick_app_api/requirements.txt
Django>=3.2.13,<3.3.0
djangorestframework>=3.13.1,<3.14.0
psycopg2-binary>=2.9.3,<2.10.0
django-cors-headers>=3.12.0,<3.13.0
Pillow>=9.1.1,<9.2.0

django-phonenumber-field>=6.1.0,<6.2.0
phonenumbers>=8.12.48,<8.13.0

celery>=5.2.6,<5.3.0
redis>=4.3.1,<4.4.0
django-celery-results>=2.3.1,<2.4.0
django-celery-beat>=2.2.1,<2.3.0

uWSGI>=2.0.20,<2.1.0
